first/backend/requirements.txt
python-socketio==5.11.4
uvicorn==0.32.0
